added support for descending 1D coordinate systems
A regular 1D grid can be ordered high-to-low as well as low-to-high. To
map to cell index space the initial array position is needed. To
determine if a point is in the grid without reading the full 1D array
the maximum value is estimated from the minimum value plus the number of
cells times their absolute spacing.